+++ .github/workflows/ci.yaml
@@ -43,4 +43,4 @@ jobs:
         uses: canonical/charm-logdump-action@main
         if: failure()
         with:
-          app: pgbouncer-k8s-operator
+          app: pgbouncer-k8s
