+++ README.md
@@ -11,7 +11,7 @@ As this charm is not yet published, you need to follow the build and deploy inst
 ## Relations
 
 - `backend:postgresql-client`
-  - Provides a relation to the corresponding [postgresql-k8s-operator charm](https://github.com/canonical/postgresql-k8s-operator).
+  - Provides a relation to the corresponding [postgresql-k8s charm](https://github.com/canonical/postgresql-k8s-operator).
   - Makes use of the [data-platform-libs DatabaseRequires library](https://github.com/canonical/data-platform-libs/blob/main/lib/charms/data_platform_libs/v0/database_provides.py).
 
 ### Legacy
