+++ src/charm.py
@@ -10,8 +10,8 @@ import os
 import socket
 from typing import Dict
 
-from charms.pgbouncer_k8s_operator.v0 import pgb
-from charms.pgbouncer_k8s_operator.v0.pgb import PgbConfig
+from charms.pgbouncer_k8s.v0 import pgb
+from charms.pgbouncer_k8s.v0.pgb import PgbConfig
 from charms.postgresql_k8s.v0.postgresql import PostgreSQL
 from ops.charm import CharmBase, ConfigChangedEvent, InstallEvent, PebbleReadyEvent
 from ops.framework import StoredState
