+++ tests/integration/relations/helpers/helpers.py
@@ -5,17 +5,17 @@
 import json
 from typing import Dict
 
-from charms.pgbouncer_k8s_operator.v0 import pgb
+from charms.pgbouncer_k8s.v0 import pgb
 from pytest_operator.plugin import OpsTest
 from tenacity import RetryError, Retrying, stop_after_delay, wait_fixed
 
-PGB = "pgbouncer-k8s-operator"
+PGB = "pgbouncer-k8s"
 
 
 def get_backend_relation(ops_test: OpsTest):
     """Gets the backend-database relation used to connect pgbouncer to the backend."""
     for rel in ops_test.model.relations:
-        if "pgbouncer-k8s-operator" in rel.endpoints and "postgresql-k8s" in rel.endpoints:
+        if "pgbouncer-k8s" in rel.endpoints and "postgresql-k8s" in rel.endpoints:
             return rel
 
     return None
@@ -75,7 +75,7 @@ async def get_userlist(ops_test: OpsTest) -> Dict[str, str]:
         "ssh",
         "--container",
         "pgbouncer",
-        "pgbouncer-k8s-operator/0",
+        "pgbouncer-k8s/0",
         "cat",
         f"{pgb.PGB_DIR}/userlist.txt",
     )
@@ -88,7 +88,7 @@ async def get_cfg(ops_test: OpsTest) -> pgb.PgbConfig:
         "ssh",
         "--container",
         "pgbouncer",
-        "pgbouncer-k8s-operator/0",
+        "pgbouncer-k8s/0",
         "cat",
         f"{pgb.PGB_DIR}/pgbouncer.ini",
     )
