+++ tests/unit/test_pgb.py
@@ -6,7 +6,7 @@ import unittest
 
 import pytest
 
-from lib.charms.pgbouncer_k8s_operator.v0 import pgb
+from lib.charms.pgbouncer_k8s.v0 import pgb
 
 DATA_DIR = "tests/unit/data"
 TEST_VALID_INI = f"{DATA_DIR}/test.ini"
