+++ tox.ini
@@ -9,7 +9,7 @@ envlist = lint, unit
 [vars]
 src_path = {toxinidir}/src/
 tst_path = {toxinidir}/tests/
-lib_path = {toxinidir}/lib/charms/pgbouncer_k8s_operator
+lib_path = {toxinidir}/lib/charms/pgbouncer_k8s
 all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
 
 [testenv]
